Enhance Warehouse Disposals Page with New Features and UI Improvements
- Introduced new data models for disposal tracking, including detailed structures for disposal receipts and returns, improving data organization.
- Added utility functions for formatting and grouping disposal records, enhancing the clarity of displayed information.
- Integrated new UI components such as dialogs and dropdowns to facilitate user interactions and improve the overall experience.
- Updated existing functions to support enhanced data handling and display logic for disposal scenarios, ensuring accurate representation of return types.

+++ supabase/migrations/20260914160000_return_leader_handovers.sql
@@ -118,7 +118,7 @@ CREATE OR REPLACE FUNCTION public.add_client_return_stock_hold(
   p_holder_id uuid,
   p_variant_id uuid,
   p_brand_id uuid,
-  p_qty integer
+  p_qty bigint
 )
 RETURNS void
 LANGUAGE plpgsql
@@ -133,7 +133,7 @@ BEGIN
   INSERT INTO public.client_return_stock_holds (
     company_id, holder_id, variant_id, brand_id, qty_on_hand
   ) VALUES (
-    p_company_id, p_holder_id, p_variant_id, p_brand_id, p_qty
+    p_company_id, p_holder_id, p_variant_id, p_brand_id, p_qty::integer
   )
   ON CONFLICT (company_id, holder_id, variant_id)
   DO UPDATE SET
@@ -149,7 +149,7 @@ CREATE OR REPLACE FUNCTION public.transfer_client_return_stock_hold(
   p_to_holder uuid,
   p_variant_id uuid,
   p_brand_id uuid,
-  p_qty integer
+  p_qty bigint
 )
 RETURNS void
 LANGUAGE plpgsql
